+++ CHANGES.rst
@@ -94,6 +94,7 @@ Internal Changes
 - Pass form class arguments to ``core.add_form_fields`` signal handlers (:pr:`6020`, thanks
   :user:`vtran99`)
 - Support Python 3.12 (:pr:`5978`)
+- Remove watchman reloader support, use watchfiles instead (:pr:`5978`)
 
 
 ----